+++ source/RTSCamera.CommandSystem/src/Logic/CommandQueueLogic.cs
@@ -68,6 +68,8 @@ namespace RTSCamera.CommandSystem.Logic
             FormationSpeedLimits.Clear();
             if (!ShouldAdjustFormationSpeed)
                 return;
+            if (CommandSystemConfig.Get().FormationSpeedSyncMode == FormationSpeedSyncMode.Disabled)
+                return;
             Dictionary<Formation, float> targetDistances = new Dictionary<Formation, float>();
             Dictionary<Formation, float> originalDurations = new Dictionary<Formation, float>();
             var maxOriginalDuration = float.MinValue;
@@ -112,54 +114,44 @@ namespace RTSCamera.CommandSystem.Logic
                 }
             }
 
-            var range = 5f;
-            foreach (var pair in targetDistances)
+            switch (CommandSystemConfig.Get().FormationSpeedSyncMode)
             {
-                var distance = pair.Value;
-                var linearSpeedLimit = distance / maxOriginalDuration;
-                var originalSpeed = MathF.Max(0.1f, pair.Key.CachedMovementSpeed);
-                var maxDistanceSpeed = GetMaxDistanceSpeed(targetDistances, pair.Value, minDistance, maxDistance, maxOriginalDuration, distanceWithMaxDuration, range);
-                var speedLimit = MathF.Lerp(maxDistanceSpeed, 0.1f, MathF.Clamp((maxDistance - distance) / range, 0f, 1f));
-                FormationSpeedLimits[pair.Key] = speedLimit;
+                case FormationSpeedSyncMode.Linear:
+                    {
+                        foreach (var pair in targetDistances)
+                        {
+                            var linearSpeedLimit = pair.Value / maxOriginalDuration;
+                            var originalSpeed = MathF.Max(0.1f, pair.Key.CachedMovementSpeed);
+                            FormationSpeedLimits[pair.Key] = linearSpeedLimit;
+                        }
+                        break;
+                    }
+                case FormationSpeedSyncMode.CatchUp:
+                    {
+                        foreach (var pair in targetDistances)
+                        {
+                            var linearSpeedLimit = MathF.Max(0.1f, pair.Value / maxOriginalDuration);
+                            var originalSpeed = MathF.Max(0.1f, pair.Key.CachedMovementSpeed);
+                            //catch up and do not wait for slower formation
+                            FormationSpeedLimits[pair.Key] = MathF.Clamp(MathF.Lerp(linearSpeedLimit, originalSpeed, (pair.Value - distanceWithMaxDuration) / (originalSpeed * 2f)), linearSpeedLimit, originalSpeed);
+                        }
+                        break;
+                    }
+                case FormationSpeedSyncMode.WaitForLastFormation:
+                    {
+                        var range = 5f;
+                        foreach (var pair in targetDistances)
+                        {
+                            var distance = pair.Value;
+                            var linearSpeedLimit = distance / maxOriginalDuration;
+                            var originalSpeed = MathF.Max(0.1f, pair.Key.CachedMovementSpeed);
+                            var maxDistanceSpeed = GetMaxDistanceSpeed(targetDistances, pair.Value, minDistance, maxDistance, maxOriginalDuration, distanceWithMaxDuration, range);
+                            var speedLimit = MathF.Lerp(maxDistanceSpeed, 0.1f, MathF.Clamp((maxDistance - distance) / range, 0f, 1f));
+                            FormationSpeedLimits[pair.Key] = speedLimit;
+                        }
+                        break;
+                    }
             }
-            //var formationDistanceList = targetDistances.ToList();
-            //formationDistanceList.Sort((KeyValuePair<Formation, float> pair1, KeyValuePair<Formation, float> pair2) =>
-            //{
-            //    var diff = pair1.Value - pair2.Value;
-            //    // descending
-            //    return diff > 0 ? -1 : diff < 0 ? 1 : 0;
-            //});
-            //var maximumSpeed = float.MaxValue;
-            //var previousDistance = float.MaxValue;
-            //for (int i = 0; i < formationDistanceList.Count; ++i)
-            //{
-            //    var formation = formationDistanceList[i].Key;
-            //    var distance = formationDistanceList[i].Value;
-            //    var linearSpeedLimit = distance / maxOriginalDuration;
-            //    var originalSpeed = MathF.Max(0.1f, formation.CachedMovementSpeed);
-            //    var minimumSpeed = linearSpeedLimit > originalSpeed ? originalSpeed : (linearSpeedLimit / originalSpeed) * linearSpeedLimit;
-            //    if (i == 0)
-            //    { 
-            //        maximumSpeed = originalSpeed;
-            //        previousDistance = distance;
-            //    }
-            //    var speedLimit = MathF.Clamp(MathF.Lerp(linearSpeedLimit, originalSpeed, MathF.Min((distance - previousDistance) / (originalSpeed * 0.1f), 1f)), minimumSpeed, maximumSpeed);
-            //    FormationSpeedLimits[formation] = speedLimit;
-            //    maximumSpeed = linearSpeedLimit;
-            //    //FormationSpeedLimits[pair.Key] = MathF.Lerp(linearSpeedLimit, originalSpeed, MathF.Max(0f, (pair.Value - distanceWithLongestDuration) / MathF.Max(1f, distanceWithLongestDuration)));
-            //    //FormationSpeedLimits[pair.Key] = linearSpeedLimit;
-            //}
-
-            //foreach (var pair in targetDistances)
-            //{
-            //    var linearSpeedLimit = pair.Value / maxOriginalDuration;
-            //    var originalSpeed = MathF.Max(0.1f, pair.Key.CachedMovementSpeed);
-            //    // catch up and wait for the last formation to arrive at target
-            //    //FormationSpeedLimits[pair.Key] = MathF.Clamp(MathF.Lerp(linearSpeedLimit, originalSpeed, (pair.Value - distanceWithLongestDuration) / (originalSpeed * 2f)), 0.1f, originalSpeed);
-            //    // catch up and do not wait for slower formation
-            //    //FormationSpeedLimits[pair.Key] = MathF.Clamp(MathF.Lerp(linearSpeedLimit, originalSpeed, (pair.Value - distanceWithLongestDuration) / (originalSpeed * 2f)), linearSpeedLimit, originalSpeed);
-            //    //FormationSpeedLimits[pair.Key] = linearSpeedLimit;
-            //}
         }
 
         private float GetMaxDistanceSpeed(Dictionary<Formation, float> targetDistances, float distance, float minDistance, float maxDistance, float maxOriginalDuration, float distanceWithMaxDuration, float range)
@@ -841,7 +833,7 @@ namespace RTSCamera.CommandSystem.Logic
                     FormationPendingOrder(formation, order);
                 }
                 order.UpdateMovementSpeed();
-                if (order.ShouldAdjustFormationSpeed && CommandSystemConfig.Get().ShouldSyncFormationSpeed && order.FormationSpeedLimits.Count > 1)
+                if (order.ShouldAdjustFormationSpeed && CommandSystemConfig.Get().FormationSpeedSyncMode != FormationSpeedSyncMode.Disabled && order.FormationSpeedLimits.Count > 1)
                 {
                     Utilities.Utility.DisplayAdjustFormationSpeedMessage(order.FormationSpeedLimits.Keys);
                 }