+++ source/RTSCamera.CommandSystem/src/Orders/VisualOrders/RTSCommandToggleFireVisualOrder.cs
@@ -2,55 +2,47 @@
 using RTSCamera.CommandSystem.Logic;
 using RTSCamera.CommandSystem.Patch;
 using System.Linq;
+using TaleWorlds.Core;
 using TaleWorlds.Localization;
 using TaleWorlds.MountAndBlade;
 using TaleWorlds.MountAndBlade.ViewModelCollection.Order.Visual;
 
 namespace RTSCamera.CommandSystem.Orders.VisualOrders
 {
-    public class RTSCommandGenericToggleVisualOrder : RTSCommandVisualOrder
+    public class RTSCommandToggleFireVisualOrder : RTSCommandVisualOrder
     {
         public static TextObject GetName(OrderType orderType)
         {
             switch (orderType)
             {
-                case OrderType.LookAtEnemy:
-                    return new TextObject("{=u8j8nN5U}Face Enemy");
-                case OrderType.LookAtDirection:
-                    return new TextObject("{=1gC25EMb}Face this Direction");
                 case OrderType.HoldFire:
                     return new TextObject("{=VyI0rimN}Holding Fire");
                 case OrderType.FireAtWill:
                     return new TextObject("{=itoYrj8d}Firing at will");
-                case OrderType.Mount:
-                    return new TextObject("{=ubTGIdcv}Mounted");
-                case OrderType.Dismount:
-                    return new TextObject("{=Ema5Vd6o}Dismounted");
-                case OrderType.AIControlOn:
-                    return new TextObject("{=zatDiaEI}Delegate Command On");
-                case OrderType.AIControlOff:
-                    return new TextObject("{=JceqNdWx}Delegate Command Off");
                 default:
                     return TextObject.GetEmpty();
             }
         }
         private readonly TextObject _positiveOrderName;
         private readonly TextObject _negativeOrderName;
+        private readonly RTSCommandToggleVolleyVisualOrder _toggleVolleyVisualOrder;
 
         public OrderType PositiveOrder { get; }
 
         public OrderType NegativeOrder { get; }
 
-        public RTSCommandGenericToggleVisualOrder(
+        public RTSCommandToggleFireVisualOrder(
             string stringId,
             OrderType positiveOrder,
-            OrderType negativeOrder)
+            OrderType negativeOrder,
+            RTSCommandToggleVolleyVisualOrder toggleVolleyVisualOrder)
             : base(stringId)
         {
             PositiveOrder = positiveOrder;
             NegativeOrder = negativeOrder;
             _positiveOrderName = GetName(positiveOrder);
             _negativeOrderName = GetName(negativeOrder);
+            _toggleVolleyVisualOrder = toggleVolleyVisualOrder;
         }
 
         public override TextObject GetName(OrderController orderController)
@@ -59,6 +51,12 @@ namespace RTSCamera.CommandSystem.Orders.VisualOrders
             {
                 case OrderState.PartiallyActive:
                 case OrderState.Active:
+                    switch (_toggleVolleyVisualOrder.GetActiveState(orderController))
+                    {
+                        case OrderState.PartiallyActive:
+                        case OrderState.Active:
+                            return _toggleVolleyVisualOrder.GetName(orderController);
+                    }
                     return _positiveOrderName;
                 default:
                     return _negativeOrderName;
@@ -79,6 +77,7 @@ namespace RTSCamera.CommandSystem.Orders.VisualOrders
             };
             orderToAdd.OrderType = GetActiveState(orderController) == OrderState.Active ? NegativeOrder : PositiveOrder;
             Patch_OrderController.LivePreviewFormationChanges.SetToggleOrder(orderToAdd.OrderType, selectedFormations);
+            Patch_OrderController.LivePreviewFormationChanges.SetVolleyEnabledOrder(false, selectedFormations);
             orderToAdd.VirtualFormationChanges = Patch_OrderController.LivePreviewFormationChanges.CollectChanges(selectedFormations);
 
             if (queueCommand)
@@ -89,6 +88,10 @@ namespace RTSCamera.CommandSystem.Orders.VisualOrders
             {
                 CommandQueueLogic.TryPendingOrder(orderToAdd.SelectedFormations, orderToAdd);
                 orderController.SetOrder(orderToAdd.OrderType);
+                foreach (var formation in orderController.SelectedFormations)
+                {
+                    CommandQueueLogic.SetFormationVolleyEnabled(formation, false);
+                }
             }
         }
 
@@ -100,7 +103,7 @@ namespace RTSCamera.CommandSystem.Orders.VisualOrders
         protected override string GetIconId()
         {
             string iconId = base.GetIconId();
-            return _lastActiveState == OrderState.Active ? iconId + "_active" : iconId;
+            return _lastActiveState == OrderState.Active ? _toggleVolleyVisualOrder.IconId : iconId;
         }
     }
 }