+++ source/RTSCamera.CommandSystem/src/Orders/VisualOrders/RTSCommandToggleVolleyVisualOrder.cs
@@ -74,7 +74,7 @@ namespace RTSCamera.CommandSystem.Orders.VisualOrders
         protected override string GetIconId()
         {
             var iconId = "order_toggle_fire";
-            return _lastActiveState == OrderState.Active ? iconId + "_active" : iconId;
+            return _lastActiveState == OrderState.Active ? iconId : iconId + "_active";
         }
     }
 }