+++ source/RTSCamera.CommandSystem/src/Patch/Patch_Formation.cs
@@ -85,6 +85,11 @@ namespace RTSCamera.CommandSystem.Patch
                         BindingFlags.Instance | BindingFlags.NonPublic),
                     prefix: new HarmonyMethod(typeof(Patch_Formation).GetMethod(
                         nameof(Prefix_ReapplyFormOrder), BindingFlags.Static | BindingFlags.Public)));
+
+                harmony.Patch(
+                    typeof(Formation).GetProperty(nameof(Formation.CalculateHasSignificantNumberOfMounted)).GetGetMethod(),
+                    prefix: new HarmonyMethod(typeof(Patch_Formation).GetMethod(
+                        nameof(Prefix_CalculateHasSignificantNumberOfMounted), BindingFlags.Static | BindingFlags.Public)));
             }
             catch (Exception e)
             {
@@ -324,6 +329,11 @@ namespace RTSCamera.CommandSystem.Patch
             __instance.FormOrder.OnApply(__instance);
             __instance.QuerySystem.Expire();
             return false;
+        }
+
+        public static bool Prefix_CalculateHasSignificantNumberOfMounted(Formation __instance, bool? ____overridenHasAnyMountedUnit, ref bool __result)
+        {
+            return ____overridenHasAnyMountedUnit.HasValue ? ____overridenHasAnyMountedUnit.Value : (double)__instance.QuerySystem.CavalryUnitRatio + (double)__instance.QuerySystem.RangedCavalryUnitRatio >= CommandSystemConfig.Get().MountedUnitsIntervalThreshold;
         }
     }
 }